# rbac: optimize rbac assigned by users query