This query was quite badly optimised. The previous query does 11 (!) joins:
```sql
SELECT DISTINCT "authentik_core_user"."id",
                "authentik_core_user"."password",
                "authentik_core_user"."last_login",
                "authentik_core_user"."username",
                "authentik_core_user"."first_name",
                "authentik_core_user"."last_name",
                "authentik_core_user"."email",
                "authentik_core_user"."is_active",
                "authentik_core_user"."date_joined",
                "authentik_core_user"."attributes",
                "authentik_core_user"."uuid",
                "authentik_core_user"."name",
                "authentik_core_user"."path",
                "authentik_core_user"."type",
                "authentik_core_user"."password_change_date",
                "authentik_core_user"."last_updated"
FROM "authentik_core_user"
LEFT OUTER JOIN "authentik_core_user_user_permissions" ON ("authentik_core_user"."id" = "authentik_core_user_user_permissions"."user_id")
LEFT OUTER JOIN "auth_permission" ON ("authentik_core_user_user_permissions"."permission_id" = "auth_permission"."id")
LEFT OUTER JOIN "django_content_type" ON ("auth_permission"."content_type_id" = "django_content_type"."id")
LEFT OUTER JOIN "guardian_userobjectpermission" ON ("authentik_core_user"."id" = "guardian_userobjectpermission"."user_id")
LEFT OUTER JOIN "auth_permission" T6 ON ("guardian_userobjectpermission"."permission_id" = T6."id")
LEFT OUTER JOIN "django_content_type" T7 ON (T6."content_type_id" = T7."id")
LEFT OUTER JOIN "authentik_core_user_ak_groups" ON ("authentik_core_user"."id" = "authentik_core_user_ak_groups"."user_id")
LEFT OUTER JOIN "authentik_core_group" ON ("authentik_core_user_ak_groups"."group_id" = "authentik_core_group"."group_uuid")
LEFT OUTER JOIN "guardian_userobjectpermission" T10 ON ("authentik_core_user"."id" = T10."user_id")
LEFT OUTER JOIN "authentik_core_user_ak_groups" T11 ON ("authentik_core_user"."id" = T11."user_id")
LEFT OUTER JOIN "authentik_core_group" T12 ON (T11."group_id" = T12."group_uuid")
WHERE ((("django_content_type"."app_label" = 'authentik_providers_proxy'
         AND "django_content_type"."model" = 'proxyprovider')
        OR (T7."app_label" = 'authentik_providers_proxy'
            AND T7."model" = 'proxyprovider')
        OR "authentik_core_group"."is_superuser")
       AND (T10."object_pk" = '1'
            OR T12."is_superuser"))
ORDER BY "authentik_core_user"."username" ASC
```
which ends up in 80 rows to be processed for a user table with 4 rows (numbers here must also take into account the quantity of rows in the joined tables, but you get the idea):
```
                                                                                                                                                                                                                                                     QUERY PLAN                                                                                                                                                                                                                                                      
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Sort  (cost=1471.26..1471.46 rows=80 width=1925) (actual time=0.252..0.255 rows=2 loops=1)
   Sort Key: authentik_core_user.username
   Sort Method: quicksort  Memory: 25kB
   ->  HashAggregate  (cost=1467.93..1468.73 rows=80 width=1925) (actual time=0.233..0.236 rows=2 loops=1)
         Group Key: authentik_core_user.username, authentik_core_user.id, authentik_core_user.password, authentik_core_user.last_login, authentik_core_user.first_name, authentik_core_user.last_name, authentik_core_user.email, authentik_core_user.is_active, authentik_core_user.date_joined, authentik_core_user.attributes, authentik_core_user.uuid, authentik_core_user.name, authentik_core_user.path, authentik_core_user.type, authentik_core_user.password_change_date, authentik_core_user.last_updated
         Batches: 1  Memory Usage: 48kB
         ->  Hash Right Join  (cost=85.38..235.49 rows=30811 width=1925) (actual time=0.175..0.191 rows=65 loops=1)
               Hash Cond: (t11.user_id = authentik_core_user.id)
               Filter: (((t10.object_pk)::text = '1'::text) OR t12.is_superuser)
               Rows Removed by Filter: 1
               ->  Hash Left Join  (cost=1.07..30.98 rows=1570 width=5) (actual time=0.015..0.016 rows=4 loops=1)
                     Hash Cond: (t11.group_id = t12.group_uuid)
                     ->  Seq Scan on authentik_core_user_ak_groups t11  (cost=0.00..25.70 rows=1570 width=20) (actual time=0.001..0.001 rows=4 loops=1)
                     ->  Hash  (cost=1.03..1.03 rows=3 width=17) (actual time=0.003..0.003 rows=4 loops=1)
                           Buckets: 1024  Batches: 1  Memory Usage: 9kB
                           ->  Seq Scan on authentik_core_group t12  (cost=0.00..1.03 rows=3 width=17) (actual time=0.001..0.001 rows=4 loops=1)
               ->  Hash  (cost=74.50..74.50 rows=785 width=1962) (actual time=0.151..0.152 rows=33 loops=1)
                     Buckets: 1024  Batches: 1  Memory Usage: 14kB
                     ->  Hash Left Join  (cost=32.34..74.50 rows=785 width=1962) (actual time=0.129..0.144 rows=33 loops=1)
                           Hash Cond: (authentik_core_user.id = t10.user_id)
                           ->  Hash Left Join  (cost=31.32..70.33 rows=785 width=1925) (actual time=0.114..0.125 rows=17 loops=1)
                                 Hash Cond: (authentik_core_user_ak_groups.group_id = authentik_core_group.group_uuid)
                                 Filter: ((((django_content_type.app_label)::text = 'authentik_providers_proxy'::text) AND ((django_content_type.model)::text = 'proxyprovider'::text)) OR (((t7.app_label)::text = 'authentik_providers_proxy'::text) AND ((t7.model)::text = 'proxyprovider'::text)) OR authentik_core_group.is_superuser)
                                 Rows Removed by Filter: 3
                                 ->  Hash Right Join  (cost=30.25..64.98 rows=1570 width=2021) (actual time=0.095..0.102 rows=20 loops=1)
                                       Hash Cond: (authentik_core_user_ak_groups.user_id = authentik_core_user.id)
                                       ->  Seq Scan on authentik_core_user_ak_groups  (cost=0.00..25.70 rows=1570 width=20) (actual time=0.004..0.004 rows=4 loops=1)
                                       ->  Hash  (cost=29.75..29.75 rows=40 width=2005) (actual time=0.082..0.083 rows=11 loops=1)
                                             Buckets: 1024  Batches: 1  Memory Usage: 11kB
                                             ->  Hash Left Join  (cost=19.03..29.75 rows=40 width=2005) (actual time=0.073..0.078 rows=11 loops=1)
                                                   Hash Cond: (authentik_core_user.id = authentik_core_user_user_permissions.user_id)
                                                   ->  Hash Left Join  (cost=9.51..20.07 rows=40 width=1965) (actual time=0.044..0.047 rows=5 loops=1)
                                                         Hash Cond: (authentik_core_user.id = guardian_userobjectpermission.user_id)
                                                         ->  Seq Scan on authentik_core_user  (cost=0.00..10.40 rows=40 width=1925) (actual time=0.005..0.006 rows=4 loops=1)
                                                         ->  Hash  (cost=9.50..9.50 rows=1 width=44) (actual time=0.027..0.028 rows=3 loops=1)
                                                               Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                                               ->  Nested Loop Left Join  (cost=0.42..9.50 rows=1 width=44) (actual time=0.018..0.025 rows=3 loops=1)
                                                                     ->  Seq Scan on guardian_userobjectpermission  (cost=0.00..1.01 rows=1 width=8) (actual time=0.005..0.005 rows=3 loops=1)
                                                                     ->  Nested Loop Left Join  (cost=0.42..8.48 rows=1 width=44) (actual time=0.005..0.006 rows=1 loops=3)
                                                                           ->  Index Scan using auth_permission_pkey on auth_permission t6  (cost=0.28..8.29 rows=1 width=8) (actual time=0.003..0.003 rows=1 loops=3)
                                                                                 Index Cond: (id = guardian_userobjectpermission.permission_id)
                                                                           ->  Index Scan using django_content_type_pkey on django_content_type t7  (cost=0.14..0.19 rows=1 width=44) (actual time=0.002..0.002 rows=1 loops=3)
                                                                                 Index Cond: (id = t6.content_type_id)
                                                   ->  Hash  (cost=9.50..9.50 rows=1 width=44) (actual time=0.020..0.020 rows=5 loops=1)
                                                         Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                                         ->  Nested Loop Left Join  (cost=0.42..9.50 rows=1 width=44) (actual time=0.011..0.018 rows=5 loops=1)
                                                               ->  Seq Scan on authentik_core_user_user_permissions  (cost=0.00..1.01 rows=1 width=8) (actual time=0.002..0.002 rows=5 loops=1)
                                                               ->  Nested Loop Left Join  (cost=0.42..8.48 rows=1 width=44) (actual time=0.003..0.003 rows=1 loops=5)
                                                                     ->  Index Scan using auth_permission_pkey on auth_permission  (cost=0.28..8.29 rows=1 width=8) (actual time=0.001..0.001 rows=1 loops=5)
                                                                           Index Cond: (id = authentik_core_user_user_permissions.permission_id)
                                                                     ->  Index Scan using django_content_type_pkey on django_content_type  (cost=0.14..0.19 rows=1 width=44) (actual time=0.001..0.001 rows=1 loops=5)
                                                                           Index Cond: (id = auth_permission.content_type_id)
                                 ->  Hash  (cost=1.03..1.03 rows=3 width=17) (actual time=0.009..0.009 rows=4 loops=1)
                                       Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                       ->  Seq Scan on authentik_core_group  (cost=0.00..1.03 rows=3 width=17) (actual time=0.005..0.006 rows=4 loops=1)
                           ->  Hash  (cost=1.01..1.01 rows=1 width=41) (actual time=0.003..0.003 rows=3 loops=1)
                                 Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                 ->  Seq Scan on guardian_userobjectpermission t10  (cost=0.00..1.01 rows=1 width=41) (actual time=0.001..0.002 rows=3 loops=1)
 Planning Time: 2.075 ms
 Execution Time: 0.457 ms
(60 rows)
```
Because the `DISTINCT` operates on all authentik_core_user columns, rows get sorted by all of the columns first, then they get filtered, and only then does the distinct happen. This causes a massive sort to happen. For big installations, this can result in the query exceeding postgres' `work_mem`, thus overflowing on disk.
We can halve the number of rows with a `DISTINCT ON (username)`:
```sql
EXPLAIN ANALYZE SELECT DISTINCT ON ("authentik_core_user"."username") "authentik_core_user"."id",
                   "authentik_core_user"."password",
                   "authentik_core_user"."last_login",
                   "authentik_core_user"."username",
                   "authentik_core_user"."first_name",
                   "authentik_core_user"."last_name",
                   "authentik_core_user"."email",
                   "authentik_core_user"."is_active",
                   "authentik_core_user"."date_joined",
                   "authentik_core_user"."attributes",
                   "authentik_core_user"."uuid",
                   "authentik_core_user"."name",
                   "authentik_core_user"."path",
                   "authentik_core_user"."type",
                   "authentik_core_user"."password_change_date",
                   "authentik_core_user"."last_updated"
FROM "authentik_core_user"
LEFT OUTER JOIN "authentik_core_user_user_permissions" ON ("authentik_core_user"."id" = "authentik_core_user_user_permissions"."user_id")
LEFT OUTER JOIN "auth_permission" ON ("authentik_core_user_user_permissions"."permission_id" = "auth_permission"."id")
LEFT OUTER JOIN "django_content_type" ON ("auth_permission"."content_type_id" = "django_content_type"."id")
LEFT OUTER JOIN "guardian_userobjectpermission" ON ("authentik_core_user"."id" = "guardian_userobjectpermission"."user_id")
LEFT OUTER JOIN "auth_permission" T6 ON ("guardian_userobjectpermission"."permission_id" = T6."id")
LEFT OUTER JOIN "django_content_type" T7 ON (T6."content_type_id" = T7."id")
LEFT OUTER JOIN "authentik_core_user_ak_groups" ON ("authentik_core_user"."id" = "authentik_core_user_ak_groups"."user_id")
LEFT OUTER JOIN "authentik_core_group" ON ("authentik_core_user_ak_groups"."group_id" = "authentik_core_group"."group_uuid")
LEFT OUTER JOIN "guardian_userobjectpermission" T10 ON ("authentik_core_user"."id" = T10."user_id")
LEFT OUTER JOIN "authentik_core_user_ak_groups" T11 ON ("authentik_core_user"."id" = T11."user_id")
LEFT OUTER JOIN "authentik_core_group" T12 ON (T11."group_id" = T12."group_uuid")
WHERE ((("django_content_type"."app_label" ='authentik_providers_proxy'
         AND "django_content_type"."model" = 'proxyprovider')
        OR (T7."app_label" = 'authentik_providers_proxy'
            AND T7."model" = 'proxyprovider')
        OR "authentik_core_group"."is_superuser")
       AND (T10."object_pk" = '1'
            OR T12."is_superuser"))
ORDER BY "authentik_core_user"."username" ASC
```
which is still not great:
```
                                                                                                                                                           QUERY PLAN                                                                                                                                                            
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=2.35..1488.26 rows=40 width=1925) (actual time=0.106..0.177 rows=2 loops=1)
   ->  Nested Loop Left Join  (cost=2.35..1411.23 rows=30811 width=1925) (actual time=0.106..0.173 rows=65 loops=1)
         Filter: (((t10.object_pk)::text = '1'::text) OR t12.is_superuser)
         Rows Removed by Filter: 1
         ->  Nested Loop Left Join  (cost=1.13..263.12 rows=785 width=1962) (actual time=0.086..0.116 rows=33 loops=1)
               Join Filter: (authentik_core_user.id = t10.user_id)
               Rows Removed by Join Filter: 19
               ->  Nested Loop Left Join  (cost=1.13..250.34 rows=785 width=1925) (actual time=0.081..0.103 rows=17 loops=1)
                     Join Filter: (authentik_core_user_ak_groups.group_id = authentik_core_group.group_uuid)
                     Rows Removed by Join Filter: 34
                     Filter: ((((django_content_type.app_label)::text = 'authentik_providers_proxy'::text) AND ((django_content_type.model)::text = 'proxyprovider'::text)) OR (((t7.app_label)::text = 'authentik_providers_proxy'::text) AND ((t7.model)::text = 'proxyprovider'::text)) OR authentik_core_group.is_superuser)
                     Rows Removed by Filter: 3
                     ->  Nested Loop Left Join  (cost=1.13..131.85 rows=1570 width=2021) (actual time=0.070..0.085 rows=20 loops=1)
                           ->  Nested Loop Left Join  (cost=0.98..68.95 rows=40 width=2005) (actual time=0.062..0.069 rows=11 loops=1)
                                 Join Filter: (authentik_core_user.id = authentik_core_user_user_permissions.user_id)
                                 Rows Removed by Join Filter: 16
                                 ->  Nested Loop Left Join  (cost=0.56..58.84 rows=40 width=1965) (actual time=0.041..0.044 rows=5 loops=1)
                                       Join Filter: (authentik_core_user.id = guardian_userobjectpermission.user_id)
                                       Rows Removed by Join Filter: 9
                                       ->  Index Scan using authentik_core_user_username_key on authentik_core_user  (cost=0.14..48.74 rows=40 width=1925) (actual time=0.011..0.012 rows=4 loops=1)
                                       ->  Materialize  (cost=0.42..9.51 rows=1 width=44) (actual time=0.006..0.007 rows=3 loops=4)
                                             ->  Nested Loop Left Join  (cost=0.42..9.50 rows=1 width=44) (actual time=0.020..0.026 rows=3 loops=1)
                                                   ->  Seq Scan on guardian_userobjectpermission  (cost=0.00..1.01 rows=1 width=8) (actual time=0.006..0.006 rows=3 loops=1)
                                                   ->  Nested Loop Left Join  (cost=0.42..8.48 rows=1 width=44) (actual time=0.005..0.006 rows=1 loops=3)
                                                         ->  Index Scan using auth_permission_pkey on auth_permission t6  (cost=0.28..8.29 rows=1 width=8) (actual time=0.003..0.003 rows=1 loops=3)
                                                               Index Cond: (id = guardian_userobjectpermission.permission_id)
                                                         ->  Index Scan using django_content_type_pkey on django_content_type t7  (cost=0.14..0.19 rows=1 width=44) (actual time=0.002..0.002 rows=1 loops=3)
                                                               Index Cond: (id = t6.content_type_id)
                                 ->  Materialize  (cost=0.42..9.51 rows=1 width=44) (actual time=0.003..0.004 rows=5 loops=5)
                                       ->  Nested Loop Left Join  (cost=0.42..9.50 rows=1 width=44) (actual time=0.010..0.017 rows=5 loops=1)
                                             ->  Seq Scan on authentik_core_user_user_permissions  (cost=0.00..1.01 rows=1 width=8) (actual time=0.002..0.002 rows=5 loops=1)
                                             ->  Nested Loop Left Join  (cost=0.42..8.48 rows=1 width=44) (actual time=0.003..0.003 rows=1 loops=5)
                                                   ->  Index Scan using auth_permission_pkey on auth_permission  (cost=0.28..8.29 rows=1 width=8) (actual time=0.001..0.001 rows=1 loops=5)
                                                         Index Cond: (id = authentik_core_user_user_permissions.permission_id)
                                                   ->  Index Scan using django_content_type_pkey on django_content_type  (cost=0.14..0.19 rows=1 width=44) (actual time=0.001..0.001 rows=1 loops=5)
                                                         Index Cond: (id = auth_permission.content_type_id)
…